+++ example/lib/example.dart
@@ -1,27 +1,13 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:mobkit_enum_generator/annotations.dart';
-
 part 'example.g.dart';
 
-@EnumSerializable(int, [
-  "succes description",
-  "inalidLogin description",
-  "lockedOut",
-  "requiresTwoFactor",
-  "passwordExpired",
-  "invalidCode"
-])
-enum ExampleEnum {
+@EnumSerializable(int, [])
+enum DailyPeriodCodeType {
   @JsonValue(1)
-  success,
-  @JsonValue(2)
-  invalidLogin,
-  @JsonValue(3)
-  lockedOut,
-  @JsonValue(4)
-  requiresTwoFactor,
-  @JsonValue(5)
-  passwordExpired,
-  @JsonValue(6)
-  invalidCode
+  morning,
+  @JsonValue(1)
+  afternoon,
+  @JsonValue(1)
+  fullDay,
 }